Additional reorganizing and renaming.
