--- data/offline/trace_entry_exit_1.json
@@ -1,1 +1,0 @@
-[]
