--- data/offline/trace_entry_exit_2.json
@@ -1,1 +1,0 @@
-[]
