--- data/offline/trace_entry_exit_3.json
@@ -1,1 +1,0 @@
-[]
