--- data/offline/trace_entry_exit_4.json
@@ -1,1 +1,0 @@
-[]
