--- data/offline/trace_send_receive.json
@@ -1,1 +1,0 @@
-[]
