--- data/online/datacfg/nwchem20180812.cfg
@@ -1,14 +1,0 @@
-# This configuration file is to be used with 
-# filternwchem.py 
-# and with the 
-# tau-nwchem 20180812 data
-# which is only partially correct
-
-[Parser]
-FileType = bp
-InputFile = datasets/tau-nwchem/20180812/tau-metrics-nwchem.bp
-OutputFile = datasets/tau-nwchem/20180812/
-ParseMode = stream
-
-[Filter]
-Threshold = 10
