--- data/online/datacfg/tauvalid20180801.cfg
@@ -1,5 +1,0 @@
-[Parser]
-FileType = bp
-InputFile = datasets/tau-validated/20180801/tau-metrics.bp
-OutputFile = datasets/tau-validated/20180801/
-ParseMode = stream
