+++ plot/run_filter.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+export PYTHONPATH=../../lib/codar/chimbuko/perf_anom:$PATH:$PYTHONPATH
+
+ARG1="$1"
+
+python3 filternwchem.py $ARG1
