+++ scripts/perfanal.cfg
@@ -5,8 +5,6 @@ ParseMode = Adios
 Method = BP
 # DATASPACES, FLEXPATH
 Parameters = verbose=3
-Timeout = timeout_sec=-1.0
-Stream = is_stream=True
 #InputFile = ../data/online/datasets/tau-nwchem/20181109/8core/tau-metrics.bp
 #InputFile = ../data/online/datasets/tau-nwchem/20181109/5core/tau-metrics.bp
 #InputFile = ../data/online/datasets/tau-nwchem/20181109/5corelong/tau-metrics.bp
@@ -42,14 +40,15 @@ VizUrl = http://0.0.0.0:80/events
 #VizUrl = http://10.2.228.192:5000/events
 VizPath = 
 DumpFile = json
+OutputFile = /home/gomfy/Workspace/Sandbox/BNL/Data/4x4x4/16core/
 
 [Debug]
-StopLoop = 2 
+StopLoop = -1 
 # If < 0 do not stop loop
 LogFile = chimbuko.log
 LogFileMode = a 
 # a, ?
-LogLevel = DEBUG 
+LogLevel = INFO 
 # CRITICAL, ERROR, WARNING, INFO, DEBUG
 Format = %(asctime)s - %(processName)s - %(threadName)s - %(filename)s - %(lineno)d - %(levelname)s - %(message)s
 # %(pathname)s
