+++ scripts/run_perfanal.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+
+#cd /Chimbuko/drivers/
+
+pwd
+
+#cd drivers/
+
+export PYTHONPATH=../lib/codar/chimbuko/perf_anom:$PATH:$PYTHONPATH
+
+ARG1="$1"
+
+python3 chimbuko.py $ARG1
